backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.2
redis==5.0.1
pydantic==2.5.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
slowapi==0.1.9
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0